+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend_module_functions.go
@@ -42,3 +42,4 @@ func (backendCore *DockerKurtosisBackend) DestroyModules(
 ) {
 	panic("Implement me")
 }
+