+++ container-engine-lib/lib/backend_interface/kurtosis_backend.go
@@ -2,6 +2,8 @@ package backend_interface
 
 import (
 	"context"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/api_container"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/engine"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/module"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/partition"
@@ -11,6 +13,9 @@ import (
 	"net"
 )
 
+// KurtosisBackend abstracts a Kurtosis backend, which will be a container engine (Docker or Kubernetes).
+// The heuristic for "do I need a method in KurtosisBackend?" here is "will I make one or more calls to
+// the underlying container engine?"
 type KurtosisBackend interface {
 	// Creates an engine with the given parameters
 	CreateEngine(
@@ -29,7 +34,7 @@ type KurtosisBackend interface {
 	// Gets engines using the given filters, returning a map of matched engines identified by their engine ID
 	GetEngines(ctx context.Context, filters *engine.EngineFilters) (map[string]*engine.Engine, error)
 
-	// Stops the engines with the given IDs
+	// Stops the engines matching the given filters
 	StopEngines(
 		ctx context.Context,
 		filters *engine.EngineFilters,
@@ -39,7 +44,7 @@ type KurtosisBackend interface {
 		resultErr error, // Represents an error with the function itself, rather than the engines
 	)
 
-	// Destroys the engines with the given IDs, regardless of if they're running or not
+	// Destroys the engines matching the given filters, regardless of if they're running or not
 	DestroyEngines(
 		ctx context.Context,
 		filters *engine.EngineFilters,
@@ -47,6 +52,91 @@ type KurtosisBackend interface {
 		successfulEngineIds map[string]bool, // "set" of engine IDs that were successfully destroyed
 		erroredEngineIds map[string]error, // "set" of engine IDs that errored when destroying, with the error
 		resultErr error, // Represents an error with the function itself, rather than the engines
+	)
+
+	// Creates an enclave with the given enclave ID
+	CreateEnclave(
+		ctx context.Context,
+		enclaveId string,
+	) (
+		*enclave.Enclave,
+		error,
+	)
+
+	// Gets enclaves matching the given filters
+	GetEnclaves(
+		ctx context.Context,
+		filters *enclave.EnclaveFilters,
+	) (
+		map[string]*enclave.Enclave,
+		error,
+	)
+
+	// TODO MAYYYYYYYBE DumpEnclaves?
+
+	// Stops enclaves matching the given filters
+	StopEnclaves(
+		ctx context.Context,
+		filters *enclave.EnclaveFilters,
+	) (
+		successfulEnclaveIds map[string]bool,
+		erroredEnclaveIds map[string]error,
+		resultErr error,
+	)
+
+	// Destroys enclaves matching the given filters
+	DestroyEnclaves(
+		ctx context.Context,
+		filters *enclave.EnclaveFilters,
+	) (
+		successfulEnclaveIds map[string]bool,
+		erroredEnclaveIds map[string]error,
+		resultErr error,
+	)
+
+	CreateAPIContainer(
+		ctx context.Context,
+		image string,
+		grpcPortId string,
+		grpcPortSpec *port_spec.PortSpec,
+		grpcProxyPortId string,
+		grpcProxyPortSpec *port_spec.PortSpec,
+		enclaveDataDirpathOnHostMachine string,	// TODO DELETE WHEN WE HAVE AN ENCLAVE DATA VOLUME!
+		envVars map[string]string,
+	) (
+		*api_container.APIContainer,
+		error,
+	)
+
+	GetAPIContainers(
+		ctx context.Context,
+		filters *api_container.APIContainerFilters,
+	) (
+	// Matching API containers, keyed by their enclave ID
+		map[string]*api_container.APIContainer,
+		error,
+	)
+
+	// Stops API containers matching the given filters
+	StopAPIContainers(
+		ctx context.Context,
+		filters *enclave.EnclaveFilters,
+	) (
+	// Successful & errored API containers are keyed by their enclave ID
+		successApiContainerIds map[string]bool,
+		erroredApiContainerIds map[string]error,
+		resultErr error,
+	)
+
+	// Stops API containers matching the given filters
+	DestroyAPIContainers(
+		ctx context.Context,
+		filters *enclave.EnclaveFilters,
+	) (
+	// Successful & errored API containers are keyed by their enclave ID
+		successApiContainerIds map[string]bool,
+		erroredApiContainerIds map[string]error,
+		resultErr error,
 	)
 
 	// Create a module from a container image with serialized params
@@ -162,4 +252,17 @@ type KurtosisBackend interface {
 	)(
 		resultErr error,
 	)
+
+	// TODO CreateRepl
+
+	// TODO AttachToRepl
+
+	// TODO GetRepls
+
+	// TODO StopRepl
+
+	// TODO DestroyRepl
+
+	// TODO RunReplExecCommand
+
 }