+++ container-engine-lib/lib/backend_interface/objects/enclave/enclave_filters.go
@@ -1,0 +1,9 @@
+package enclave
+
+type EnclaveFilters struct {
+	// Disjunctive set of enclave IDs to operate on
+	IDs map[string]bool
+
+	// Disjunctive set of enclave statuses to operate on
+	Statuses map[EnclaveStatus]bool
+}