+++ container-engine-lib/docs/changelog.md
@@ -3,6 +3,19 @@
 * Added `UserService` methods, `Modules` methods and `CreateRepartition` method in `KurtosisBackend` interface
 * Studding out methods for `UserService`, `Modules` and `CreateRepartition` into Docker implementation
 
+# 0.10.1
+### Features
+* Added API container CRUD stubs to `KurtosisBackend`
+
+# 0.10.0
+### Features
+* Created a generic `ContainerStatus` object in the `KurtosisBackend` API for representing the state of containers
+* Added enclave CRUD commands to `KurtosisBackend`
+
+### Breaking Changes
+* The `Engine` object's status will now be a `ContainerStatus`, rather than `EngineStatus`
+    * Users should migrate to the new version
+
 # 0.9.1
 ### Fixes
 * Fixed container `removing` state erroneously counting as running